test(Schemas): Move schema checking to tests
