+++ package.json
@@ -9,13 +9,13 @@
   "types": "dist/types.d.ts",
   "scripts": {
     "lint": "prettier --write './**/*.{js,json,md,ts,yaml}' && eslint './**/*.{js,ts}' --fix",
-    "test": "gulp test",
-    "build": "npm run build:jsonschema && npm run build:ts && npm run build:py",
-    "build:jsonschema": "ts-node src/schema.ts",
+    "build": "npm run build:schema && npm run build:ts && npm run build:py",
+    "build:schema": "ts-node src/schema.ts",
     "build:jsonld": "gulp jsonld",
     "build:ts": "ts-node src/typescript.ts",
     "build:py": "ts-node src/python.ts",
     "build:r": "ts-node src/r.ts",
+    "test": "jest",
     "docs": "npm run docs:readme && npm run docs:build",
     "docs:readme": "markdown-toc -i --maxdepth=4 README.md",
     "docs:build": "ts-node scripts/docs && scripts/docs.sh",
@@ -36,6 +36,7 @@
     "@stencila/dev-config": "^1.1.2",
     "@stencila/encoda": "^0.59.1",
     "@types/fs-extra": "^8.0.0",
+    "@types/jest": "^24.0.15",
     "@types/js-yaml": "^3.12.1",
     "@types/toposort": "^2.0.1",
     "ajv": "^6.10.2",
@@ -44,6 +45,8 @@
     "globby": "^10.0.1",
     "gulp": "^4.0.2",
     "husky": "^3.0.0",
+    "jest": "^24.8.0",
+    "jest-file-snapshot": "^0.3.6",
     "js-yaml": "^3.13.1",
     "json-schema": "^0.2.3",
     "json-schema-to-typescript": "^6.1.3",
@@ -52,6 +55,7 @@
     "tempy": "^0.3.0",
     "through2": "^3.0.1",
     "toposort": "^2.0.2",
+    "ts-jest": "^24.0.2",
     "ts-node": "^8.3.0",
     "typescript": "^3.5.3",
     "vscode-json-languageservice": "^3.2.0"
@@ -80,6 +84,10 @@
     "built",
     "dist"
   ],
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node"
+  },
   "husky": {
     "hooks": {
       "pre-commit": "pretty-quick --staged",
