+++ src/schema.ts
@@ -22,12 +22,14 @@ if (module.parent === null) build()
  */
 export async function build(): Promise<void> {
   // Asynchronously read all the schema definition YAML files into a map of objects
-  const files = await globby('*.schema.yaml', {cwd: SCHEMA_SOURCE_DIR})
+  const files = await globby('*.schema.yaml', { cwd: SCHEMA_SOURCE_DIR })
   const schemas = new Map<string, Schema>(
     await Promise.all(
       files.map(
         async (file: string): Promise<[string, Schema]> => {
-          const schema = yaml.safeLoad(await fs.readFile(path.join(SCHEMA_SOURCE_DIR, file), 'utf-8'))
+          const schema = yaml.safeLoad(
+            await fs.readFile(path.join(SCHEMA_SOURCE_DIR, file), 'utf-8')
+          )
           const title = schema.title
           if (title === undefined)
             throw new Error(`Schema title is required in source file: ${file}`)
@@ -44,7 +46,7 @@ export async function build(): Promise<void> {
   await fs.ensureDir('built')
   await Promise.all(
     Array.from(schemas.entries()).map(async ([title, schema]) => {
-      const destPath = path.join(SCHEMA_DEST_DIR,title +  '.schema.json')
+      const destPath = path.join(SCHEMA_DEST_DIR, title + '.schema.json')
       await fs.writeJSON(destPath, schema, { spaces: 2 })
     })
   )
@@ -158,7 +160,11 @@ function processSchema(schemas: Map<string, Schema>, schema: Schema): void {
     const walk = (node: Schema): void => {
       if (typeof node !== 'object') return
       for (const [key, child] of Object.entries(node)) {
-        if (key === '$ref' && typeof child === 'string' && !child.endsWith('.schema.json'))
+        if (
+          key === '$ref' &&
+          typeof child === 'string' &&
+          !child.endsWith('.schema.json')
+        )
           node[key] = child + '.schema.json'
         walk(child)
       }
